3.2.5

+++ src/functions/tver_functions.ps1
@@ -44,23 +44,25 @@ function Get-VideoLinksFromKeyword {
 		specialMainLinks = [System.Collections.Generic.List[string]]::new()
 		specialLinks     = [System.Collections.Generic.List[string]]::new()
 	}
-	$key = $keyword.split(' ')[0].split("`t")[0].Split('/')[0]
-	$tverID = Remove-Comment(($keyword.Replace("$key/", '')).Trim())
-	if ($key -eq $tverID) { $key = 'episodes' }
-	if (($tverID -eq 'sitemap') -or ($tverID -eq 'toppage')) { $key = $tverID }
+	if ($keyword.IndexOf('/') -gt 0) { 
+		$key = $keyword.split(' ')[0].split("`t")[0].Split('/')[0]
+		$tverID = Remove-Comment(($keyword.Replace("$key/", '')).Trim())
+	} else { $key = '' ; $tverID = '' }
+	if (($keyword -eq 'sitemap') -or ($keyword -eq 'toppage')) { $key = $keyword }
 	Invoke-StatisticsCheck -Operation 'search' -TVerType $key -TVerID $tverID
 	switch ($key) {
-		'series' { $linkCollection.seriesLinks.Add($tverID); continue }
-		'talents' { $linkCollection.talentLinks.Add($tverID); continue }
-		'tag' { $result = Get-LinkFromTag $tverID; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'new' { $result = Get-LinkFromNew $tverID; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'ranking' { $result = Get-LinkFromRanking $tverID; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'toppage' { $result = Get-LinkFromTopPage; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'sitemap' { $result = Get-LinkFromSiteMap; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'mypage' { $result = Get-LinkFromMyPage $tverID; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-		'episodes' { $linkCollection.episodeLinks.Add(('https://tver.jp/{0}/{1}' -f $key, $tverID)); continue }
-		default { $result = Get-LinkFromFreeKeyword $keyword; $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result }
+		'series' { $linkCollection.seriesLinks.Add($tverID) ; continue }
+		'talents' { $linkCollection.talentLinks.Add($tverID) ; continue }
+		'episodes' { $linkCollection.episodeLinks.Add(('https://tver.jp/{0}/{1}' -f $key, $tverID)) ; continue }
+		'mypage' { $result = Get-LinkFromMyPage $tverID ; continue }
+		'tag' { $result = Get-LinkFromTag $tverID ; continue }
+		'ranking' { $result = Get-LinkFromRanking $tverID ; continue }
+		'new' { $result = Get-LinkFromNew $tverID ; continue }
+		'toppage' { $result = Get-LinkFromTopPage ; continue }
+		'sitemap' { $result = Get-LinkFromSiteMap ; continue }
+		default { $result = Get-LinkFromFreeKeyword $keyword }
 	}
+	if (Test-Path Variable:result) { $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result }
 	while (($linkCollection.specialMainLinks.Count -ne 0) -or ($linkCollection.specialLinks.Count -ne 0) -or ($linkCollection.talentLinks.Count -ne 0) -or ($linkCollection.seriesLinks.Count -ne 0) -or ($linkCollection.seasonLinks.Count -ne 0)) {
 		if ($linkCollection.specialMainLinks) {
 			$linkCollection = Convert-Buffer -TverIDs $linkCollection.specialMainLinks -TverIDType 'Special Main' -LinkCollection $linkCollection
@@ -98,7 +100,8 @@ function ProcessSearchResults {
 		[Parameter(Mandatory = $true)][String]$baseURL,
 		[Parameter(Mandatory = $false)][String]$type,
 		[Parameter(Mandatory = $false)][String]$keyword,
-		[Parameter(Mandatory = $false)][String]$requireData
+		[Parameter(Mandatory = $false)][String]$requireData,
+		[Parameter(Mandatory = $false)][Boolean]$loginRequired
 	)
 	Write-Debug ('{0}' -f $MyInvocation.MyCommand.Name)
 	$episodeLinks = [System.Collections.Generic.List[string]]::new()
@@ -109,19 +112,10 @@ function ProcessSearchResults {
 	$specialLinks = [System.Collections.Generic.List[string]]::new()
 	#URLの整形
 	$sid = $script:myMemberSID
-	$uid = $script:platformUID
-	$token = $script:platformToken
-	if ($sid) {
-		# TVerIDにログインして使う場合
-		$callSearchURL = '{0}?member_sid={1}' -f $baseURL, $sid
-	} else {
-		# TVerIDを匿名で使う場合
-		if ($requireData -and $script:myPlatformUID -ne '' -and $script:myPlatformToken -ne '') {
-			$uid = $script:myPlatformUID
-			$token = $script:myPlatformToken
-		}
-		$callSearchURL = '{0}?platform_uid={1}&platform_token={2}' -f $baseURL, $uid, $token
-	}
+	if (($script:myPlatformUID -ne '') -and ($script:myPlatformToken -ne '')) { $uid = $script:myPlatformUID ; $token = $script:myPlatformToken }
+	else { $uid = $script:platformUID ; $token = $script:platformToken }
+	if ($loginRequired) { $callSearchURL = '{0}?member_sid={1}' -f $baseURL, $sid }						# TVerIDにログインして使う場合
+	else { $callSearchURL = '{0}?platform_uid={1}&platform_token={2}' -f $baseURL, $uid, $token }		# TVerIDを匿名で使う場合
 	switch ($type) {
 		'keyword' { if ($keyword) { $callSearchURL += "&keyword=$keyword" } }
 		'mypage' { if ($requireData) { $callSearchURL += "&require_data=$requireData" } }
@@ -146,11 +140,11 @@ function ProcessSearchResults {
 				'talent' { $talentLinks.Add($searchResult.Content.Id) ; continue }
 				'special' {
 					if ($type -eq 'specialmain') { $specialLinks.Add($searchResult.Content.Id) }
-					else { $episodeLinks = & { $episodeLinks; (Get-LinkFromSpecialDetailID $searchResult.Content.Id) } }
+					else { $episodeLinks = & { $episodeLinks ; (Get-LinkFromSpecialDetailID $searchResult.Content.Id) } }
 					continue
 				}
-				'specialMain' { $specialMainLinks.Add($searchResult.Content.Id); continue }
-				default { $episodeLinks = & { $episodeLinks; ('https://tver.jp/{0}/{1}' -f $searchResult.Type, $searchResult.Content.Id) } }
+				'specialMain' { $specialMainLinks.Add($searchResult.Content.Id) ; continue }
+				default { $episodeLinks = & { $episodeLinks ; ('https://tver.jp/{0}/{1}' -f $searchResult.Type, $searchResult.Content.Id) } }
 			}
 		}
 	} catch {
@@ -158,7 +152,7 @@ function ProcessSearchResults {
 			Write-Warning ('⚠️ HTTP接続がタイムアウトしました。スキップして次のリンクを処理します。')
 			Write-Warning ('　　{0}, {1}, {2}' -f $type, $keyword, $requireData)
 		} elseif ($_.Exception.Message.Contains('Response status code does not indicate success:')) {
-			Write-Warning ('⚠️ HTTP接続が失敗しました。スキップして次のリンクを処理します。 - {0}' -f $_.Exception.Message.Replace('Response status code does not indicate success:', ''))
+			Write-Warning ('⚠️ HTTP接続が失敗しました。スキップして次のリンクを処理します。 - {0}' -f $_.Exception.Message)
 			Write-Warning ('　　{0}, {1}, {2}' -f $type, $keyword, $requireData)
 		} else {
 			Write-Warning ('⚠️ エラーが発生しました。スキップして次のリンクを処理します。 - {0}' -f $_.Exception.Message)
@@ -321,10 +315,11 @@ function Get-LinkFromTopPage {
 	}
 	$callSearchBaseURL = 'https://platform-api.tver.jp/service/api/v1/callHome'
 	$callSearchURL = ('{0}?platform_uid={1}&platform_token={2}' -f $callSearchBaseURL, $script:platformUID, $script:platformToken)
-	$searchResults = Invoke-RestMethod -Uri $callSearchURL -Method 'GET' -Headers $script:requestHeader -TimeoutSec $script:timeoutSec
+	try { $searchResults = Invoke-RestMethod -Uri $callSearchURL -Method 'GET' -Headers $script:requestHeader -TimeoutSec $script:timeoutSec }
+	catch { Write-Warning 'トップページを取得できませんでした' ; return $linkCollection }
 	foreach ($component in $searchResults.Result.Components) {
 		switch ($component.Type) {
-			{ $_ -in @('horizontal', 'ranking', 'talents', 'billboard', 'episodeRanking', 'newer', 'ender', 'talent', 'special', 'specialContent', 'topics', 'spikeRanking') } {
+			{ $_ -in @('horizontal', 'richHorizontal', 'ranking', 'talents', 'billboard', 'episodeRanking', 'newer', 'ender', 'talent', 'special', 'specialContent', 'topics', 'spikeRanking') } {
 				$contents = if ($component.Type -eq 'topics') { $component.Contents.Content.Content } else { $component.Contents }
 				foreach ($content in $contents) {
 					if ($content.Type -eq 'live') { continue }
@@ -335,12 +330,13 @@ function Get-LinkFromTopPage {
 						'talent' { $linkCollection.talentLinks.Add($content.Content.Id) ; continue }
 						'specialMain' { $linkCollection.specialMainLinks.Add($content.Content.Id) ; continue }
 						'special' { $linkCollection.specialLinks.Add($content.Content.Id) ; continue }
-						default { Write-Warning ('⚠️ 未知のパターンです。 - {0}/{1}' -f $content.Type, $content.Content.Id) }
+						default { Write-Warning ('⚠️ 未知のコンテンツタイプです。 - {0}/{1}' -f $content.Type, $content.Content.Id) }
 					}
 				}
+				continue
 			}
-			{ $_ -in @('banner', 'resume', 'favorite') } { }
-			default { Write-Warning "⚠️ 未知のパターンです。 - $($component.Type)" }
+			{ $_ -in @('banner', 'resume', 'favorite') ; continue } {}
+			default { Write-Warning "⚠️ 未知のコンポーネントタイプです。 - $($component.Type)" }
 		}
 	}
 	return $linkCollection
@@ -364,13 +360,16 @@ function Get-LinkFromSiteMap {
 		specialLinks     = [System.Collections.Generic.List[string]]::new()
 	}
 	$callSearchURL = 'https://tver.jp/sitemap.xml'
-	$searchResultsRaw = Invoke-RestMethod -Uri $callSearchURL -Method 'GET' -Headers $script:requestHeader -TimeoutSec $script:timeoutSec
+	try { $searchResultsRaw = Invoke-RestMethod -Uri $callSearchURL -Method 'GET' -Headers $script:requestHeader -TimeoutSec $script:timeoutSec }
+	catch { Write-Warning 'サイトマップを取得できませんでした' ; return $linkCollection }
 	$searchResults = $searchResultsRaw.urlset.url.loc | Sort-Object -Unique
 	foreach ($url in $searchResults) {
 		try {
-			$url = $url.Replace('https://tver.jp/', '')
-			$url = $url -split '/'
-			$tverID = @{ type = $url[0] ; id = $url[1] }
+			$url = $url.Replace('https://tver.jp/', '') -split '/'
+			$tverID = @{
+				type = $url[0]
+				id   = $url[1]
+			}
 		} catch { $tverID = @{ type = $null ; id = $null } }
 		if ($tverID.id) {
 			switch ( $tverID.type) {
@@ -416,15 +415,15 @@ function Get-LinkFromMyPage {
 	[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseOutputTypeCorrectly', '')]
 	Param ([Parameter(Mandatory = $true, ValueFromPipeline = $true)][String]$page)
 	Write-Debug ('{0}' -f $MyInvocation.MyCommand.Name)
-	$baseURLPrefix = if ($script:myMemberSID) { 'https://member-api.tver.jp' } else { 'https://platform-api.tver.jp' }
+	$baseURLPrefix = if ($script:myMemberSID) { 'https://member-api.tver.jp' ; $loginRequired = $true } else { 'https://platform-api.tver.jp' ; $loginRequired = $false }
 	switch ($page) {
-		'fav' { $baseURL = ('{0}/service/api/v2/callMylistDetail/{1}' -f $baseURLPrefix, (ConvertTo-UnixTime (Get-Date))); $requireData = 'mylist'; continue }
-		'later' { $baseURL = ('{0}/service/api/v2/callMyLater' -f $baseURLPrefix); $requireData = 'later'; continue }
-		'resume' { $baseURL = ('{0}/service/api/v2/callMyResume' -f $baseURLPrefix); $requireData = 'resume'; continue }
-		'favorite' { $baseURL = ('{0}/service/api/v2/callMyFavorite' -f $baseURLPrefix); $requireData = 'mylist'; continue }
+		'fav' { $baseURL = ('{0}/service/api/v2/callMylistDetail/{1}' -f $baseURLPrefix, (ConvertTo-UnixTime (Get-Date))) ; $requireData = 'mylist' ; continue }
+		'later' { $baseURL = ('{0}/service/api/v2/callMyLater' -f $baseURLPrefix) ; $requireData = 'later' ; continue }
+		'resume' { $baseURL = ('{0}/service/api/v2/callMyResume' -f $baseURLPrefix) ; $requireData = 'resume' ; continue }
+		'favorite' { $baseURL = ('{0}/service/api/v2/callMyFavorite' -f $baseURLPrefix) ; $requireData = 'mylist' ; continue }
 		default { Write-Warning "⚠️ 未知のパターンです。 - mypage/$page" }
 	}
-	$tverIDs = ProcessSearchResults -baseURL $baseURL -Type 'mypage' -RequireData $requireData
+	$tverIDs = ProcessSearchResults -baseURL $baseURL -Type 'mypage' -RequireData $requireData -LoginRequired $loginRequired
 	return $tverIDs
 	Remove-Variable -Name page, baseURL, requireData, tverIDs -ErrorAction SilentlyContinue
 }
@@ -433,7 +432,7 @@ function Get-LinkFromMyPage {
 #TVerIDの整理
 #----------------------------------------------------------------------
 function Update-LinkCollection {
-	param (
+	Param (
 		[PSCustomObject]$linkCollection,
 		[PSCustomObject]$result
 	)
@@ -445,7 +444,7 @@ function Update-LinkCollection {
 	if ($result.specialMainLinks) { $linkCollection.specialMainLinks = [System.Collections.Generic.List[string]](& { $linkCollection.specialMainLinks ; $result.specialMainLinks }) }
 	if ($result.specialLinks) { $linkCollection.specialLinks = [System.Collections.Generic.List[string]](& { $linkCollection.specialLinks ; $result.specialLinks }) }
 	return $linkCollection
-	Remove-Variable -Name linkCollection, result -ErrorAction SilentlyContinue
+	Remove-Variable -Name linkTypes, linkType, linkCollection, result -ErrorAction SilentlyContinue
 }
 
 #----------------------------------------------------------------------
@@ -455,7 +454,7 @@ function Convert-Buffer {
 	[CmdletBinding()]
 	[OutputType([PSCustomObject])]
 	[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseOutputTypeCorrectly', '')]
-	param(
+	Param (
 		[Parameter(Mandatory = $false)][Object[]]$tverIDs,
 		[Parameter(Mandatory = $true)][ValidateSet('Special Main', 'Special Detail', 'Talent', 'Season', 'Series')][string]$tverIDType,
 		[Parameter(Mandatory = $true)][OutputType([PSCustomObject])]$linkCollection
@@ -464,18 +463,17 @@ function Convert-Buffer {
 	if ($tverIDs) {
 		foreach ($tverID in ($tverIDs | Sort-Object -Unique)) {
 			Write-Information ('{0} - {1} {2} からEpisodeを抽出中...' -f (Get-Date), $tverIDType, $tverID)
-			switch ($tverIDType) {
-				'Special Main' { $result = Get-LinkFromSpecialMainID($tverID); $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-				'Special Detail' { $result = Get-LinkFromSpecialDetailID($tverID); $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-				'Talent' { $result = Get-LinkFromTalentID($tverID); $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-				'Series' { $result = Get-LinkFromSeriesID($tverID); $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
-				'Season' { $result = Get-LinkFromSeasonID($tverID); $linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result; continue }
+			$result = switch ($tverIDType) {
+				'Special Main' { Get-LinkFromSpecialMainID($tverID) ; continue }
+				'Special Detail' { Get-LinkFromSpecialDetailID($tverID) ; continue }
+				'Talent' { Get-LinkFromTalentID($tverID) ; continue }
+				'Series' { Get-LinkFromSeriesID($tverID) ; continue }
+				'Season' { Get-LinkFromSeasonID($tverID) ; continue }
 			}
+			$linkCollection = Update-LinkCollection -linkCollection $linkCollection -result $result 
 		}
-		if ($linkCollection.episodeLinks) { $linkCollection.episodeLinks = $linkCollection.episodeLinks | Sort-Object -Unique }
-		if ($linkCollection.seasonLinks) { $linkCollection.seasonLinks = $linkCollection.seasonLinks | Sort-Object -Unique }
-		if ($linkCollection.seriesLinks) { $linkCollection.seriesLinks = $linkCollection.seriesLinks | Sort-Object -Unique }
-		if ($linkCollection.specialLinks) { $linkCollection.specialLinks = $linkCollection.specialLinks | Sort-Object -Unique }
+		$linkTypes = @('episodeLinks', 'seasonLinks', 'seriesLinks', 'specialLinks')
+		foreach ($linkType in $linkTypes) { if ($linkCollection.$linkType) { $linkCollection.$linkType = $linkCollection.$linkType | Sort-Object -Unique } }
 	}
 	return $linkCollection
 	Remove-Variable -Name tverIDs, tverIDType, linkCollection, tverID, result -ErrorAction SilentlyContinue
